there is no long at all in the X specification
SVN revision: 65090

+++ legacy/ecore/src/lib/ecore_x/xcb/ecore_xcb_window_prop.c
@@ -113,7 +113,7 @@ ecore_x_window_prop_card32_list_get(Ecore_X_Window win,
                }
              data = xcb_get_property_value(reply);
              for (i = 0; i < num; i++)
-               val[i] = ((unsigned long *)data)[i];
+               val[i] = ((unsigned int *)data)[i];
              *list = val;
           }
      }
